modified showing of percents

+++ MoneyBuddi/src/controller/AddExpenseServlet.java
@@ -56,7 +56,7 @@ public class AddExpenseServlet extends HttpServlet {
 				double amount=Double.parseDouble(amountAsString);
 				long categoryId=Long.parseLong(request.getParameter("categoryId"));
 				Category category=CategoryDAO.getInstance().getCategoryByID(categoryId);
-				Integer accountId=(Integer)request.getSession().getAttribute("accountId");
+				long accountId=(long)request.getSession().getAttribute("accountId");
 				Account account=AccountDao.getInstance().getAccountById(accountId);
 				Currency currency=account.getCurrency();
